简单调整alder

+++ libHDiffPatch/HDiff/private_diff/limit_mem_diff/adler_roll.h
@@ -69,12 +69,12 @@ extern "C" {
 
 #define  __private_fast_adler_roll(uint_t,half_bit, \
                                   adler,blockSize,out_data,in_data){ \
-    uint32_t in_v=(uint32_t)in_data*in_data;    \
+    uint32_t in_v=((uint32_t)in_data)*in_data;  \
     uint32_t out_v=(uint32_t)out_data*out_data; \
     uint_t sum=adler>>half_bit;  \
     adler= adler + in_v - out_v; \
-    sum  = sum + adler - ADLER_INITIAL-(uint32_t)blockSize*out_v; \
-    return (adler&(((uint_t)1<<half_bit)-1)) | (sum<<half_bit);   \
+    sum  = sum + adler - ADLER_INITIAL-(uint32_t)(((uint32_t)blockSize)*out_v); \
+    return (adler&(((uint_t)1<<half_bit)-1)) | (sum<<half_bit); \
 }
 
 uint32_t adler32_append(uint32_t adler,const adler_data_t* pdata,size_t n);
